renamed computer parsers to laptop parsers (for clearity)
